✨ update:시간표시함수elapsedTime를 common폴더로 이동 및 다른 js에 적용.
