+++ front/js/post/comments.js
@@ -1,4 +1,5 @@
 import { API_BASE_URL } from '../../config/config.js';
+import { elapsedTime } from '../common/elapsed-time.js';
 
 const submitCommentButton = document.getElementById('submit-comment');
 const commentContentInput = document.getElementById('comment-content');
@@ -39,7 +40,7 @@ function addCommentToList(comment) {
   const commentItem = document.createElement('li');
   commentItem.innerHTML = `
                 <p>${comment.nickname} 
-                | 작성일: ${new Date(comment.createdAt).toLocaleString()} 
+                | 작성일: ${elapsedTime(comment.createdAt)} 
                 | 좋아요: ${comment.likes || 0} 
                 | 싫어요: ${comment.dislikes || 0}
                 </p>
@@ -61,7 +62,7 @@ function addCommentToList(comment) {
       const recommentItem = document.createElement('li');
       recommentItem.innerHTML = `
                       <p>${recomment.nickname}
-                      | 작성일: ${new Date(recomment.createdAt).toLocaleString()}
+                      | 작성일: ${elapsedTime(recomment.createdAt)}
                       | 좋아요: ${recomment.likes}
                       | 싫어요: ${recomment.dislikes}</p>
                       <p>${recomment.content}</p>
